+++ R/benchmark_grid.R
@@ -98,10 +98,10 @@ benchmark_grid = function(tasks, learners, resamplings, param_values = NULL, pai
       task = tasks[[i]]
       resampling = resamplings[[i]]
       if (!resamplings[[i]]$is_instantiated) {
-        error_input("Resampling #%i ('%s' for task '%s') is not instantiated", i, resampling$id, task$id)
+        error_config("Resampling #%i ('%s' for task '%s') is not instantiated", i, resampling$id, task$id)
       }
       if (resampling$task_row_hash != task$row_hash) {
-        error_input(
+        error_config(
           "Resampling #%i ('%s' for task '%s') is not instantiated on the corresponding task",
           i,
           resampling$id,
@@ -118,7 +118,7 @@ benchmark_grid = function(tasks, learners, resamplings, param_values = NULL, pai
 
     if (any(is_instantiated) && !all(is_instantiated)) {
       # prevent that some resamplings are instantiated and others are not
-      error_input("All resamplings must be instantiated, or none at all")
+      error_config("All resamplings must be instantiated, or none at all")
     } else if (all(is_instantiated)) {
       # check that all row ids of the resamplings are present in the tasks
       pwalk(grid, function(task, resampling) {
@@ -126,7 +126,7 @@ benchmark_grid = function(tasks, learners, resamplings, param_values = NULL, pai
           !is.null(resamplings[[resampling]]$task_row_hash) &&
             resamplings[[resampling]]$task_row_hash != tasks[[task]]$row_hash
         ) {
-          error_input(
+          error_config(
             "Resampling '%s' is not instantiated on task '%s'",
             resamplings[[resampling]]$id,
             tasks[[task]]$id